backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
python-docx==1.1.2
httpx==0.28.1
python-dotenv==1.1.1
